Renamed tests in UpperCamelCase
