+++ FinancialMathematics/phpunit-tests/InterestFormatTest.php
@@ -1,6 +1,6 @@
 <?php
 
-require 'test-constants.php';
+require 'TestConstants.php';
 require_once $class_directory . 'class-ct1-interest-format.php';
 class CT1_Interest_Format_Test extends PHPUnit_Framework_TestCase
 {
