+++ FinancialMathematics/phpunit-tests/MortgageTest.php
@@ -1,6 +1,6 @@
 <?php
 
-require 'test-constants.php';
+require 'TestConstants.php';
 require_once $class_directory . 'class-ct1-mortgage.php';
 class CT1_Mortgage_Test extends PHPUnit_Framework_TestCase
 {
