+++ FinancialMathematics/phpunit-tests/XMLTest.php
@@ -1,6 +1,6 @@
 <?php
 
-require 'test-constants.php';
+require 'TestConstants.php';
 require_once $class_directory . 'class-ct1-xml.php';
 class CT1_XML_Test extends PHPUnit_Framework_TestCase
 {
